fixing front issues (log in problem)

+++ backendVscode/src/main/java/com/example/auth/dto/CommentDTO.java
@@ -17,6 +17,6 @@ public class CommentDTO {
     String url;
     String type;
     @NotNull
-    int postId;
+    Long postId;
 
 }
